+++ src/iampypsa/couplers/__init__.py
@@ -1,0 +1,8 @@
+"""Coupler interface and IAM-specific backends."""
+
+from __future__ import annotations
+
+from iampypsa.couplers.base import CouplingAdapter
+from iampypsa.couplers.remind import RemindGdxAdapter, RemindIamcAdapter
+
+__all__ = ["CouplingAdapter", "RemindGdxAdapter", "RemindIamcAdapter"]
